bouncer-test/requirements.txt
fastapi
uvicorn
httpx
pydantic
google-generativeai